Fix: Entity 이름 수정 및 패키지 이동
