--- src/main/java/com/cocos/cocos/common/response/message/ApiMessage.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.common.response.message;
-
-import org.springframework.http.HttpStatus;
-
-public interface ApiMessage {
-    HttpStatus getHttpStatus();
-    int getCode();
-    String getMessage();
-}
