--- src/main/java/com/cocos/cocos/common/response/message/FailMessage.java
@@ -1,67 +1,0 @@
-package com.cocos.cocos.common.response.message;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-@Getter
-@AllArgsConstructor
-public enum FailMessage implements ApiMessage{
-
-    /**
-     * 400
-     */
-    BAD_REQUEST(HttpStatus.BAD_REQUEST, 40000, "잘못된 요청입니다. "),
-    BAD_REQUEST_REQUEST_BODY_VALID(HttpStatus.BAD_REQUEST, 40001,"request body 검증 실패입니다. "),
-    BAD_REQUEST_MISSING_PARAM(HttpStatus.BAD_REQUEST,40002, "필수 param이 없습니다. "),
-    BAD_REQUEST_METHOD_ARGUMENT_TYPE(HttpStatus.BAD_REQUEST, 40003,"메서드 인자타입이 잘못되었습니다. "),
-    BAD_REQUEST_NOT_READABLE(HttpStatus.BAD_REQUEST, 40004,"json 오류 혹은 reqeust body 필드 오류 입니다. "),
-
-    /**
-     * 401
-     */
-    UNAUTHORIZED(HttpStatus.UNAUTHORIZED, 40100,"인증이 필요합니다. "),
-
-    /**
-     * 403
-     */
-    FORBIDDEN(HttpStatus.FORBIDDEN, 40300,"권한이 없습니다. "),
-
-    /**
-     * 404
-     */
-    NOT_FOUND(HttpStatus.NOT_FOUND, 40400,"리소스를 찾을 수 없습니다. "),
-    NOT_FOUND_ENTITY(HttpStatus.NOT_FOUND, 40401,"대상을 찾을 수 없습니다. "),
-    NOT_FOUND_API(HttpStatus.NOT_FOUND,40402, "잘못된 API입니다. "),
-
-    /**
-     * 405
-     */
-    METHOD_NOT_ALLOWED(HttpStatus.METHOD_NOT_ALLOWED, 40500,"잘못된 HTTP method 요청입니다."),
-
-    /**
-     * 409
-     */
-    CONFLICT(HttpStatus.CONFLICT, 40900,"데이터 충돌이 발생했습니다. "),
-    INTEGRITY_CONFLICT(HttpStatus.CONFLICT, 40901,"데이터 무결성 위반입니다."),
-
-    /**
-     * 422
-     */
-    UNPROCESSABLE_ENTITY(HttpStatus.UNPROCESSABLE_ENTITY, 42200,"처리할 수 없는 요청입니다."),
-
-    /**
-     * 500
-     */
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, 50000,"서버 내부 오류가 발생했습니다. "),
-
-    /**
-     * 503
-     */
-    SERVICE_UNAVAILABLE(HttpStatus.SERVICE_UNAVAILABLE, 50300,"현재 서비스를 사용할 수 없습니다. ");
-
-
-    private final HttpStatus httpStatus;
-    private final int code;
-    private final String message;
-}
