--- src/main/java/com/cocos/cocos/db/category/repository/CategoryRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.category.repository;
-
-import com.cocos.cocos.db.category.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
