+++ src/main/java/com/cocos/cocos/db/comment/entity/SubComment.java
@@ -1,0 +1,27 @@
+package com.cocos.cocos.db.comment.entity;
+
+import com.cocos.cocos.db.BaseTime;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@Table(name = "sub_comment")
+public class SubComment extends BaseTime {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
+    private Long id;
+
+    @Column(name = "content", nullable = false)
+    private String content;
+
+    @Column(name = "member_id", nullable = false)
+    private Long memberId;
+
+    @Column(name = "comment_id", nullable = false)
+    private Long commentId;
+}
