+++ src/main/java/com/cocos/cocos/db/comment/repository/SubCommentRepository.java
@@ -1,0 +1,9 @@
+package com.cocos.cocos.db.comment.repository;
+
+import com.cocos.cocos.db.comment.entity.SubComment;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface SubCommentRepository extends JpaRepository<SubComment, Long> {
+}
