--- src/main/java/com/cocos/cocos/db/heart/entity/Heart.java
@@ -1,24 +1,0 @@
-package com.cocos.cocos.db.heart.entity;
-
-import com.cocos.cocos.db.BaseTime;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "heart")
-public class Heart extends BaseTime {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "member_id", nullable = false)
-    private Long memberId;
-
-    @Column(name = "post_id", nullable = false)
-    private Long postId;
-}
