--- src/main/java/com/cocos/cocos/db/heart/repository/HeartRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.heart.repository;
-
-import com.cocos.cocos.db.heart.entity.Heart;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface HeartRepository extends JpaRepository<Heart, Long> {
-}
