--- src/main/java/com/cocos/cocos/db/image/repository/ImageRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.image.repository;
-
-import com.cocos.cocos.db.image.entity.Image;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ImageRepository extends JpaRepository<Image, Long> {
-}
