+++ src/main/java/com/cocos/cocos/db/member/repository/MemberHospitalRepository.java
@@ -1,0 +1,9 @@
+package com.cocos.cocos.db.member.repository;
+
+import com.cocos.cocos.db.member.entity.MemberHospital;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberHospitalRepository extends JpaRepository<MemberHospital, Long> {
+}
