--- src/main/java/com/cocos/cocos/db/memberhospital/entity/MemberHospital.java
@@ -1,24 +1,0 @@
-package com.cocos.cocos.db.memberhospital.entity;
-
-import com.cocos.cocos.db.BaseTime;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "member_hospital")
-public class MemberHospital extends BaseTime {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "hospital_id", nullable = false)
-    private Long hospitalId;
-
-    @Column(name = "member_id", nullable = false)
-    private Long memberId;
-}
