--- src/main/java/com/cocos/cocos/db/memberhospital/repository/MemberHospitalRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.memberhospital.repository;
-
-import com.cocos.cocos.db.memberhospital.entity.MemberHospital;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberHospitalRepository extends JpaRepository<MemberHospital, Long> {
-}
