+++ src/main/java/com/cocos/cocos/db/pet/entity/PetSymptom.java
@@ -1,0 +1,23 @@
+package com.cocos.cocos.db.pet.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@Table(name = "pet_symptom")
+public class PetSymptom {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
+    private Long id;
+
+    @Column(name = "pet_id", nullable = false)
+    private Long petId;
+
+    @Column(name = "symptom_id", nullable = false)
+    private Long symptomId;
+}
