--- src/main/java/com/cocos/cocos/db/petdisease/entity/PetDisease.java
@@ -1,23 +1,0 @@
-package com.cocos.cocos.db.petdisease.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "pet_disease")
-public class PetDisease {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "pet_id", nullable = false)
-    private Long petId;
-
-    @Column(name = "disease_id", nullable = false)
-    private Long diseaseId;
-}
