--- src/main/java/com/cocos/cocos/db/petdisease/repository/PetDiseaseRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.petdisease.repository;
-
-import com.cocos.cocos.db.petdisease.entity.PetDisease;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PetDiseaseRepository extends JpaRepository<PetDisease, Long> {
-}
