+++ src/main/java/com/cocos/cocos/db/post/entity/PostCategory.java
@@ -1,0 +1,23 @@
+package com.cocos.cocos.db.post.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+@Table(name = "post_category")
+public class PostCategory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id", nullable = false)
+    private Long id;
+
+    @Column(name = "name", nullable = false)
+    private String name;
+
+    @Column(name = "image", nullable = false)
+    private String image;
+}
