+++ src/main/java/com/cocos/cocos/db/post/repository/PostImageRepository.java
@@ -1,0 +1,9 @@
+package com.cocos.cocos.db.post.repository;
+
+import com.cocos.cocos.db.post.entity.PostImage;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PostImageRepository extends JpaRepository<PostImage, Long> {
+}
