--- src/main/java/com/cocos/cocos/db/subcomment/entity/SubComment.java
@@ -1,27 +1,0 @@
-package com.cocos.cocos.db.subcomment.entity;
-
-import com.cocos.cocos.db.BaseTime;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "sub_comment")
-public class SubComment extends BaseTime {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "content", nullable = false)
-    private String content;
-
-    @Column(name = "member_id", nullable = false)
-    private Long memberId;
-
-    @Column(name = "comment_id", nullable = false)
-    private Long commentId;
-}
