--- src/main/java/com/cocos/cocos/db/tag/repository/TagRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.tag.repository;
-
-import com.cocos.cocos.db.tag.entity.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
-}
