--- src/main/java/com/cocos/cocos/db/token/entity/Token.java
@@ -1,26 +1,0 @@
-package com.cocos.cocos.db.token.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-@Table(name = "token")
-public class Token {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "refresh_token", nullable = false)
-    private String refreshToken;
-
-    @Column(name = "kakao_refresh_token", nullable = false)
-    private String KakaoRefreshToken;
-
-    @Column(name = "member_id", nullable = false)
-    private Long memberId;
-}
