--- src/main/java/com/cocos/cocos/db/token/repository/TokenRepository.java
@@ -1,9 +1,0 @@
-package com.cocos.cocos.db.token.repository;
-
-import com.cocos.cocos.db.token.entity.Token;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TokenRepository extends JpaRepository<Token, Long> {
-}
